Move billing form to separate page and update landing with diagram
